Adding explicit numeic casting to sys.sp_databases_view
Error found while upgrading babelfishpg_tsql extension

running sql: ALTER EXTENSION "babelfishpg_tsql" UPDATE, error: ERROR: operator is not unique: bigint / numeric Hint: Could not choose a best candidate operator. You might need to add explicit type casts.

Solution
Following the error hint (as below) resolves the issue, so adding explicit NUMERIC casting to sys.sp_databases_view:

select pg_relation_size('sys.babelfish_namespace_ext')::numeric/1024.0;

+++ contrib/babelfishpg_tsql/sql/upgrades/babelfishpg_tsql--2.4.0--2.5.0.sql
@@ -7,8 +7,8 @@ SELECT set_config('search_path', 'sys, '||current_setting('search_path'), false)
 CREATE OR REPLACE VIEW sys.sp_databases_view AS
 	SELECT CAST(database_name AS sys.SYSNAME),
 	-- DATABASE_SIZE returns a NULL value for databases larger than 2.15 TB
-	CASE WHEN (sum(table_size)/1024.0) > 2.15 * 1024.0 * 1024.0 * 1024.0 THEN NULL
-		ELSE CAST((sum(table_size)/1024.0) AS int) END as database_size,
+	CASE WHEN (sum(table_size)::NUMERIC/1024.0) > 2.15 * 1024.0 * 1024.0 * 1024.0 THEN NULL
+		ELSE CAST((sum(table_size)::NUMERIC/1024.0) AS int) END as database_size,
 	CAST(NULL AS sys.VARCHAR(254)) as remarks
 	FROM (
 		SELECT pg_catalog.pg_namespace.oid as schema_oid,
